crawler_service/requirements.txt
Django==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
newspaper3k==0.2.8
feedparser==6.0.10
beautifulsoup4==4.12.3